Add files via upload
Created a project on the problem statement (Smart and Effective real time management of street parking). We have integrated software and hardware for this purpose. We should be placing Arduino boards with ultrasonic sensors at parking spaces and then connecting it with ESP8266 chips. After the configuration, we shall be embedding the IP addresses of Arduino boards in their respective parking blocks in the website. This board will be helping us in showing empty spaces and the ultrasonic sensor will be continuously checking for the distance between the car and the board which in case increased by more than 3 meters i.e the case of theft , it will send an alert to the website.
Now, when a user books a space he/she  can turn on the Arduino board and hence, turning on the sensors too.